tests: split the ssh harness out of e2e.rs
No behavior change.
